\name{MRprollim-package}
\alias{MRprollim-package}
\alias{MRprollim}
\docType{package}
\title{
  MR-PROLLIM
}
\description{
  Mendelian randomization with pleiotropy-robust log-linear models adapted for binary outcomes and potential horizontal pleiotropy.
}
\details{
  This section should provide a more detailed overview of how to use the
  package, including the most important functions.
}
\author{
Mintao Li, Jinzhu Jia, email optional.

Maintainer: Mintao Li <limintao@pku.edu.cn>
}
\references{
  This optional section can contain literature or other references for
  background information.
}
\keyword{ package }
\seealso{
  Optional links to other man pages
}
\examples{
  \dontrun{
     ## Optional simple examples of the most important functions
     ## These can be in \dontrun{} and \donttest{} blocks.   
  }
}
